try a cargo test style
