+++ tools/httpbintest/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "http3bin-client"
+name = "httpbintest"
 version = "0.1.0"
 authors = ["Lucas Pardue <lucaspardue.24.7@gmail.com>"]
 edition = "2018"
@@ -7,9 +7,9 @@ edition = "2018"
 [dependencies]
 docopt = "1"
 env_logger = "0.6"
-mio = "0"
+mio = "0.6"
 url = "1"
-log = "0"
+log = "0.4"
 ring = "0.16"
 serde = "1.0"
 serde_json = "1.0"
